Populated left panel with multiple frames:  1. connect frame on the top needs user input for SN of PM100USB  2. Sensor frame detects connected sensor and outputs sensor info  3. Device frame has controls for PM100USB once it establishes valid     connection
Need to improve layout, it is disorganized and not clean

Populated right panel for power readout, currently it does not output
anything. Later we need to readout power continuously.

Also need to think about power units when in W mode. Preferrably, the
units should automatically change to appropriate order of magnitude,
e.g, when in e-6 range it should display uW etc.